James/requirements.txt
discord.py==1.7.3
python-dotenv
pytest
dpytest